+++ engine/graphics/private/gpu_buffer_uploader.hpp
@@ -18,7 +18,7 @@ public:
         rhi_buffer* buffer,
         const void* data,
         std::size_t size,
-        std::uint32_t offset,
+        std::size_t offset,
         rhi_pipeline_stage_flags stages,
         rhi_access_flags access);
     void record(rhi_command* command);
