+++ engine/plugins/vulkan/include/vk_context.hpp
@@ -88,17 +88,17 @@ public:
         return m_physical_device_properties;
     }
 
-    std::size_t get_frame_count() const noexcept
+    std::uint32_t get_frame_count() const noexcept
     {
         return m_frame_count;
     }
 
-    std::size_t get_frame_resource_count() const noexcept
+    std::uint32_t get_frame_resource_count() const noexcept
     {
         return m_frame_resource_count;
     }
 
-    std::size_t get_frame_resource_index() const noexcept
+    std::uint32_t get_frame_resource_index() const noexcept
     {
         return m_frame_resource_index;
     }
@@ -142,9 +142,9 @@ private:
 
     VmaAllocator m_vma_allocator{VK_NULL_HANDLE};
 
-    std::size_t m_frame_count{0};
-    std::size_t m_frame_resource_count{0};
-    std::size_t m_frame_resource_index{0};
+    std::uint32_t m_frame_count{0};
+    std::uint32_t m_frame_resource_count{0};
+    std::uint32_t m_frame_resource_index{0};
 
     std::mutex m_mutex;
