+++ engine/plugins/vulkan/include/vk_rhi.hpp
@@ -25,17 +25,17 @@ public:
     void begin_frame() override;
     void end_frame() override;
 
-    std::size_t get_frame_count() const noexcept override
+    std::uint32_t get_frame_count() const noexcept override
     {
         return m_context->get_frame_count();
     }
 
-    std::size_t get_frame_resource_count() const noexcept override
+    std::uint32_t get_frame_resource_count() const noexcept override
     {
         return m_context->get_frame_resource_count();
     }
 
-    std::size_t get_frame_resource_index() const noexcept override
+    std::uint32_t get_frame_resource_index() const noexcept override
     {
         return m_context->get_frame_resource_index();
     }
